# blake256: Full redesign.