--- a/crypto/blake256/README.md
+++ b/crypto/blake256/README.md
@@ -1,7 +1,0 @@
-Package blake256
-================
-
-Package blake256 implements BLAKE-256 and BLAKE-224 hash functions (SHA-3
-candidate).
-
-Originally from `github.com/teknico/blake256`.