--- a/crypto/blake256/internal/_asm/go.mod
+++ b/crypto/blake256/internal/_asm/go.mod
@@ -1,0 +1,10 @@
+module github.com/decred/dcrd/crypto/blake256/internal/_asm
+
+go 1.18
+
+require github.com/mmcloughlin/avo v0.6.0
+
+require (
+	golang.org/x/mod v0.14.0 // indirect
+	golang.org/x/tools v0.16.1 // indirect
+)