The current `blake256` implementation has served well enough over the years, however, the underlying implementation it was based on has a few less than desirable qualities.  For example, it was not designed in a way that is conducive to avoiding memory allocations nor exposing new functionality because it only returns a standard library interface.  In addition, the core compression code is not very well optimized, particularly for smaller input sizes which are extremely common.

Consequently, this consists of a series of commits which rewrite the package from scratch to address the shortcomings while significantly optimizing it along the way to include adding asm optimizations that take advantage of SIMD vector extensions (SSE2, SSE4.1, and AVX) on amd64 processors that support them.  It does not require any dependencies beyond the standard library.

The design philosophy of the new implementation places a strong emphasis correctness, readability, and efficiency while also aiming to provide an ergonomic API.

In order to avoid needing a new major module version, care has been taken to expose all existing public functions, methods, and constants in addition to the new concrete types and methods that expose the new capabilities.  Naturally, callers using the old API will still be subject to its limitations.  However, they will still gain the speed benefits of the new better optimized implementations which are used under the hood.

Several steps have been taken to ease the review process.  For example, the first commit only implements the pure Go code along with some basic infrastructure to support implementing the aforementioned asm optimizations while the optimizations themselves are implemented in subsequent commits.

Next, in order to significantly ease the review and implementation of the assembly code that takes advantage of the amd64 SIMD vector extensions, it is generated with Go via avo and includes extensive comments.  This means only the Go code that generates the asm needs to be asserted for correctness versus needing to review thousands of lines of pure asm code.  It also means most of the complexity is handled in Go at compile time and Go controls structures can be used without the associated overhead in the generated asm since the result is completely unrolled.

Since the overall structure is the same for all specialized implementations given they ultimately must all perform the same process, the generation code itself is abstracted to simplify new additions.  Other than typical function declaration and related setup, each specialized implementation only is required to populate the state of the `blocksCompressor` struct with the relevant registers and functions from a new instance and call its generate method.

The generation code is implemented via a separate internal module to ensure the specific version of avo used to generate the code is pinned and therefore entirely reproducible without adding an otherwise unnecessary dependency to the main public module.

The assembly code can be generating by running `go generate` in the `_asm` directory.  It can then be compared against the generated code included in the commits.

### Overview of some of the key highlights of the new implementation

- Enables zero allocations for rolling hashes
- Concrete rolling hashers (for BLAKE-224 and BLAKE-256) with convenience methods for directly writing unsigned integers of various sizes encoded as big or little endian as well as single bytes
- Ability to save and restore the intermediate state (also often called the midstate):
  - Same process state saving with a simple instance copy
  - Cross process state saving via serialization
- Provides asm implementations accelerated by SSE2, SSE4.1, and AVX that are dynamically selected based on processor support
  - Supports the `purego` build tag to disable all asm code if desired
  - Supports environment variables to selectively disable extensions
- Hashing with salt
- Comprehensive test coverage with 100% branch coverage
- Full suite of benchmarks for analyzing performance
- Supporting documentation
- Example usage for various scenarios
- No dependencies other than the standard library

### Performance Comparisons

As the following sections show, the new implementation is notably faster when only using pure Go, and significantly faster when SIMD vector extensions are available.  All benchmarks are the result of feeding `benchstat` 10 iterations.

#### Performance Summary of Various Implementations

The following results demonstrate the performance of hashing various amounts of data for both small and larger inputs with the `Sum256` method.

Operation        |   Pure Go   |    SSE2     |   SSE4.1    |     AVX     | Allocs / Op
-----------------|-------------|-------------|-------------|-------------|------------
`Sum256` (32b)   | 190ns ± 1%  |  170ns ± 1% |  138ns ± 0% |  137ns ± 1% | 0
`Sum256` (64b)   | 342ns ± 0%  |  308ns ± 0% |  237ns ± 1% |  236ns ± 1% | 0
`Sum256` (1KiB)  | 2.71µs ± 1% | 2.43µs ± 1% | 1.91µs ± 1% | 1.90µs ± 1% | 0
`Sum256` (8KiB)  | 20.2µs ± 1% | 18.3µs ± 0% | 14.3µs ± 0% | 14.3µs ± 0% | 0
`Sum256` (16KiB) | 40.8µs ± 1% | 36.5µs ± 0% | 28.5µs ± 0% | 28.5µs ± 0% | 0

Operation        |   Pure Go    |     SSE2     |    SSE4.1    |    AVX
-----------------|--------------|--------------|--------------|-------------
`Sum256` (32b)   | 168MB/s ± 1% | 188MB/s ± 1% | 232MB/s ± 0% | 234MB/s ± 1%
`Sum256` (64b)   | 187MB/s ± 0% | 208MB/s ± 0% | 270MB/s ± 1% | 271MB/s ± 1%
`Sum256` (1KiB)  | 378MB/s ± 1% | 421MB/s ± 1% | 536MB/s ± 1% | 539MB/s ± 1%
`Sum256` (8KiB)  | 405MB/s ± 1% | 448MB/s ± 0% | 573MB/s ± 0% | 573MB/s ± 0%
`Sum256` (16KiB) | 402MB/s ± 1% | 449MB/s ± 0% | 575MB/s ± 0% | 575MB/s ± 0%

#### Pure Go New vs Existing Comparison

The following is a sample showing the delta of the BLAKE-256 performance of inputs of various sizes using the new Pure Go implementation versus the old implementation it replaces.  Note that this puts the old implementation in the best light possible by avoiding the interface that forces allocations, so, in actual practice, there is typically an even larger difference in performance. 

Operation         | Old Time/Op | New Time/Op | Delta
------------------|-------------|-------------|--------
Hasher256 (32b)   |  225ns ± 1% |  190ns ± 1% | -15.56%
Hasher256 (64b)   |  407ns ± 2% |  342ns ± 0% | -15.97%
Hasher256 (1KiB)  | 2.83µs ± 2% | 2.71µs ± 1% | -4.32%
Hasher256 (8KiB)  | 20.9µs ± 2% | 20.2µs ± 1% | -3.35%
Hasher256 (16KiB) | 42.5µs ± 3% | 40.8µs ± 1% | -4.00%

Operation         | Old Speed    | New Speed    | Delta
------------------|--------------|--------------| -------
Hasher256 (32b)   | 143MB/s ± 2% | 168MB/s ± 1% | +17.48%
Hasher256 (64b)   | 157MB/s ± 2% | 187MB/s ± 0% | +19.11%
Hasher256 (1KiB)  | 362MB/s ± 2% | 378MB/s ± 1% |  +4.51%
Hasher256 (8KiB)  | 391MB/s ± 2% | 405MB/s ± 1% |  +3.58%
Hasher256 (16KiB) | 386MB/s ± 3% | 402MB/s ± 1% |  +4.14%

#### AVX Optimized vs Existing Comparison

Operation         | Old Time/Op | New Time/Op | Delta
------------------|-------------|-------------|--------
Hasher256 (32b)   |  225ns ± 1% |  139ns ± 1% | -38.05%
Hasher256 (64b)   |  407ns ± 2% |  241ns ± 1% | -40.75%
Hasher256 (1KiB)  | 2.83µs ± 2% | 1.92µs ± 1% | -32.25%
Hasher256 (8KiB)  | 20.9µs ± 2% | 14.4µs ± 1% | -31.25%
Hasher256 (16KiB) | 42.5µs ± 3% | 28.5µs ± 1% | -32.87%

Operation         | Old Speed    | New Speed    | Delta
------------------|--------------|--------------| -------
Hasher256 (32b)   | 143MB/s ± 2% | 230MB/s ± 1% | +61.40%
Hasher256 (64b)   | 157MB/s ± 2% | 265MB/s ± 1% | +68.79%
Hasher256 (1KiB)  | 362MB/s ± 2% | 534MB/s ± 1% | +47.59%
Hasher256 (8KiB)  | 391MB/s ± 2% | 569MB/s ± 1% | +45.45%
Hasher256 (16KiB) | 386MB/s ± 3% | 574MB/s ± 1% | +48.92%