@teknico 2024-07-15
Lately I was considering an SIMD blake256 implementation using avo, interesting coincidence after five years from last time we worked on this. I’ll have a closer look at this tomorrow.

How come no AVX2 version using the added “gather” instructions? +70% is great, but I was hoping for something closer to the theoretical 4x speed up.

Impressive work, as usual.