@davecgh 2024-07-15
@teknico Thanks.  I made a comment in the generation code about AVX2:

```
// NOTE: Various attempts to optimize using the larger 256-bit registers
// provided by AVX2 were made, but since only 4 columns can be computed in
// parallel, it turns out that the extra overhead of shuffling data around
// offsets any gains made by the few places that the larger registers are
// able to speed up.  That includes things such as converting the message to
// big endian using 2x256-bit registers and freeing up registers by packing
// more data into the larger registers and then making use of the extra
// freed up registers to cache the results of xoring the message and
// constants to reuse in final rounds where they are the same.
```

I also tried with `VPGATHERDD`, but it was ultimately actually slower due to the memory accesses versus the register-only approach I ended up implementing.

All of the infrastructure is in place for AVX2 including the feature detection, so if you're able to figure out how to make it faster with AVX2 using some things I wasn't able to think of, I'd be more than happy to review it and get it merged.