@jrick 2024-07-15
There seems to be a small performance regression on arm64.

Old benchmarks:

```
Benchmark1K-8             216348              8105 ns/op         126.34 MB/s
Benchmark8K-8              19833             60506 ns/op         135.39 MB/s
Benchmark64-8            1000000              1057 ns/op          60.57 MB/s
Benchmark1KNoAlloc-8      214888              8078 ns/op         126.76 MB/s
Benchmark8KNoAlloc-8       27813             60422 ns/op         135.58 MB/s
Benchmark64NoAlloc-8     1000000              1070 ns/op          59.80 MB/s
```

New:

```
BenchmarkHasher256/32b-8                 2196926               546.2 ns/op        58.59 MB/s           0 B/op          0 allocs/op
BenchmarkHasher256/64b-8                 1000000              1035 ns/op          61.85 MB/s           0 B/op          0 allocs/op
BenchmarkHasher256/1KiB-8                 143847              8339 ns/op         122.80 MB/s           0 B/op          0 allocs/op
BenchmarkHasher256/8KiB-8                  19076             62893 ns/op         130.25 MB/s           0 B/op          0 allocs/op
BenchmarkHasher256/16KiB-8                  9573            125251 ns/op         130.81 MB/s           0 B/op          0 allocs/op
```