@jrick 2024-07-15
(btw, it would have been nice if one of the first commits in this series added benchmarks with the same names as implemented later, to make benchcmp easier to check)