@davecgh 2024-07-15
@jrick: Thanks for comparing.  Can you check if the `g` and `binary.BigEndian.Uint32` functions in `internal/compress/blocks_generic.go` are being inlined?