@jrick 2024-07-15
```
x13s$ benchstat old.txt new.txt                               
goos: openbsd
goarch: arm64
pkg: github.com/decred/dcrd/crypto/blake256
               │   old.txt    │       new.txt       │
               │    sec/op    │    sec/op     vs base   │
1K-8             8.103µ ±  0%
8K-8             60.48µ ±  0%
64-8             1.092µ ±  1%
1KNoAlloc-8      8.078µ ± 31%
8KNoAlloc-8      60.41µ ±  0%
64NoAlloc-8      1.062µ ± 31%
Sum256/32b-8                    557.2n ± 29%
Sum256/64b-8                    1.049µ ±  2%
Sum256/1KiB-8                   8.346µ ±  0%
Sum256/8KiB-8                   62.86µ ±  0%
Sum256/16KiB-8                  125.2µ ±  0%
geomean          8.075µ         8.257µ        ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
² ratios must be >0 to compute geomean

               │    old.txt    │       new.txt        │
               │      B/s      │      B/s       vs base   │
1K-8             120.5Mi ±  0%
8K-8             129.2Mi ±  0%
64-8             55.89Mi ±  1%
1KNoAlloc-8      120.9Mi ± 45%
8KNoAlloc-8      129.3Mi ±  0%
64NoAlloc-8      57.49Mi ± 46%
Sum256/32b-8                     54.77Mi ± 41%
Sum256/64b-8                     58.21Mi ±  2%
Sum256/1KiB-8                    117.0Mi ±  0%
Sum256/8KiB-8                    124.3Mi ±  0%
Sum256/16KiB-8                   124.8Mi ±  0%
geomean          95.99Mi         89.64Mi        ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
² ratios must be >0 to compute geomean

               │   new.txt    │
               │     B/op     │
Sum256/32b-8     0.000 ± 0%
Sum256/64b-8     0.000 ± 0%
Sum256/1KiB-8    0.000 ± 0%
Sum256/8KiB-8    0.000 ± 0%
Sum256/16KiB-8   0.000 ± 0%
geomean                     ¹
¹ summaries must be >0 to compute geomean

               │   new.txt    │
               │  allocs/op   │
Sum256/32b-8     0.000 ± 0%
Sum256/64b-8     0.000 ± 0%
Sum256/1KiB-8    0.000 ± 0%
Sum256/8KiB-8    0.000 ± 0%
Sum256/16KiB-8   0.000 ± 0%
geomean                     ¹
¹ summaries must be >0 to compute geomean
```