@matthawkins90 2024-07-16
Latest changes benchmarked on an M2 mac:

```sh
./compare.sh
go: downloading github.com/decred/dcrd/crypto/blake256 v1.0.0
go: downloading github.com/davecgh/dcrd/crypto/blake256 v1.0.1-0.20240716021806-5ca4d50cad45
goos: darwin
goarch: arm64
pkg: blake256bench
                │   old.txt   │              new.txt               │
                │   sec/op    │   sec/op     vs base               │
Sum256/32b-12     184.1n ± 0%   174.0n ± 1%  -5.49% (p=0.000 n=10)
Sum256/64b-12     333.9n ± 0%   328.0n ± 0%  -1.77% (p=0.000 n=10)
Sum256/1KiB-12    2.570µ ± 0%   2.621µ ± 3%       ~ (p=0.101 n=10)
Sum256/8KiB-12    19.27µ ± 1%   19.82µ ± 0%  +2.85% (p=0.000 n=10)
Sum256/16KiB-12   38.32µ ± 0%   39.44µ ± 1%  +2.91% (p=0.000 n=10)
geomean           2.591µ        2.592µ       +0.04%

                │   old.txt    │               new.txt               │
                │     B/s      │     B/s       vs base               │
Sum256/32b-12     165.8Mi ± 0%   175.4Mi ± 1%  +5.79% (p=0.000 n=10)
Sum256/64b-12     182.8Mi ± 0%   186.1Mi ± 0%  +1.79% (p=0.000 n=10)
Sum256/1KiB-12    379.9Mi ± 0%   372.6Mi ± 3%       ~ (p=0.105 n=10)
Sum256/8KiB-12    405.4Mi ± 0%   394.1Mi ± 0%  -2.78% (p=0.000 n=10)
Sum256/16KiB-12   407.7Mi ± 0%   396.2Mi ± 1%  -2.82% (p=0.000 n=10)
geomean           285.7Mi        285.6Mi       -0.04%

                │   old.txt    │               new.txt               │
                │     B/op     │    B/op     vs base                 │
Sum256/32b-12     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum256/64b-12     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum256/1KiB-12    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum256/8KiB-12    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum256/16KiB-12   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                      ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                │   old.txt    │               new.txt               │
                │  allocs/op   │ allocs/op   vs base                 │
Sum256/32b-12     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum256/64b-12     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum256/1KiB-12    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum256/8KiB-12    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Sum256/16KiB-12   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                      ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```